Query: Re-organize SelectExpression
Part of #17337

This commit doesn't make any functional change. Just moved code around for better organization and made certain private functions local static functions
